+++ tx-models/src/main/java/qualitynotification/base/response/QualityNotificationResponse.java
@@ -20,13 +20,19 @@ package qualitynotification.base.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.Size;
+import lombok.AllArgsConstructor;
 import lombok.Data;
+import lombok.RequiredArgsConstructor;
 import lombok.experimental.SuperBuilder;
 
+import java.util.Comparator;
 import java.util.List;
+import java.util.Optional;
 
 @Data
 @SuperBuilder
+@AllArgsConstructor
+@RequiredArgsConstructor
 public class QualityNotificationResponse {
 
     @Schema(example = "66", maxLength = 255)
@@ -81,5 +87,10 @@ public class QualityNotificationResponse {
     private String targetDate;
 
     private List<QualityNotificationMessageResponse> messages;
+
+
+    public Optional<QualityNotificationMessageResponse> latestNotification() {
+        return messages.stream().max(Comparator.comparing(QualityNotificationMessageResponse::getCreated)).stream().findFirst();
+    }
 
 }